moved form validation of cameracreation to controller
